# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Operator HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/push/devices:
    post:
      operationId: registerPushDevice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPushDeviceRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushDeviceEnvelope'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Register (upsert) a phone's Expo push token
      tags:
      - push
  /api/v1/push/devices/{token}:
    delete:
      operationId: unregisterPushDevice
      parameters:
      - description: Expo push token (URL-encoded) identifying the device.
        in: path
        name: token
        required: true
        schema:
          description: Expo push token (URL-encoded) identifying the device.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnregisterPushDeviceResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Unregister a phone's Expo push token
      tags:
      - push
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    PushDeviceEnvelope:
      properties:
        device:
          $ref: '#/components/schemas/PushDeviceResponse'
      required:
      - device
      type: object
    PushDeviceResponse:
      properties:
        createdAt:
          format: date-time
          type: string
        deviceName:
          type: string
        lastSeenAt:
          format: date-time
          type: string
        platform:
          type: string
        token:
          type: string
      required:
      - token
      - createdAt
      - lastSeenAt
      type: object
    RegisterPushDeviceRequest:
      properties:
        deviceName:
          description: Human-friendly device label.
          type: string
        platform:
          description: Device platform.
          enum:
          - ios
          - android
          type: string
        token:
          description: Expo push token, e.g. ExponentPushToken[...].
          type: string
      required:
      - token
      type: object
...
    UnregisterPushDeviceResponse:
      properties:
        deleted:
          type: boolean
        token:
          type: string
      required:
      - token
      - deleted
      type: object
...